+++ src/app/app-routing.module.ts
@@ -1,23 +1,20 @@
+import { MetalwebsComponent } from './webshop/metalwebs/metalwebs.component';
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { NagelplattenBasketComponent } from './home/nagelplatten-basket/nagelplatten-basket.component';
 import { NagelplattenComponent } from './webshop/nagelplatten/nagelplatten.component';
 
 const routes: Routes = [
   {
     path: '',
-    component: NagelplattenComponent
+    redirectTo: '/nagelplatten',
+    pathMatch: 'full'
   },
   {
     path: 'nagelplatten',
     component: NagelplattenComponent
   },
-  { path: 'basket',
-    component: NagelplattenBasketComponent
-  },
-  {
-    path: 'dashboard',
-    loadChildren: '../app/dashboard/dashboard.module#DashboardModule'
+  { path: 'metalwebs',
+    component: MetalwebsComponent
   }
 ];
