--- src/app/dashboard/dashboard-home/dashboard-home.component.html
@@ -1,3 +1,0 @@
-<p>
-  dashboard-home works!
-</p>
