--- src/app/dashboard/dashboard-home/dashboard-home.component.ts
@@ -1,16 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-import { RouterModule } from '@angular/router';
-
-@Component({
-  selector: 'app-dashboard-home',
-  templateUrl: './dashboard-home.component.html',
-  styleUrls: ['./dashboard-home.component.scss']
-})
-export class DashboardHomeComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
