--- src/app/dashboard/dashboard-routing.module.ts
@@ -1,16 +1,0 @@
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-import { DashboardHomeComponent } from './dashboard-home/dashboard-home.component';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: DashboardHomeComponent
-  }
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule]
-})
-export class DashboardRoutingModule { }
