--- src/app/dashboard/dashboard.module.ts
@@ -1,15 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { RouterModule } from '@angular/router';
-
-import { DashboardRoutingModule } from './dashboard-routing.module';
-import { DashboardHomeComponent } from './dashboard-home/dashboard-home.component';
-
-@NgModule({
-  imports: [
-    CommonModule,
-    DashboardRoutingModule
-  ],
-  declarations: [DashboardHomeComponent]
-})
-export class DashboardModule { }
