+++ src/app/sidenav/sidenav/sidenav.component.html
@@ -1,8 +1,8 @@
 <div class="main-sidebar"><div class="sidebar">
     <div class="vertical-menu">
-        <a routerLink='/nagelplatten' routerLinkActive="active-link" [routerLinkActiveOptions]="{exact:
+        <a routerLink='/nagelplatten' routerLinkActive="active-link-1" [routerLinkActiveOptions]="{exact:
           true}">Bestellung Nagelplatten</a>
-        <a routerLink='/basket' routerLinkActive="active-link" [routerLinkActiveOptions]="{exact:
+        <a routerLink='/metalwebs' routerLinkActive="active-link" [routerLinkActiveOptions]="{exact:
           true}">Bestellung Metal Webs</a>
         <a href="#">Bestellung Binderwinkel</a>
         <a href="#">AGB Bestellung</a>
