+++ src/app/sidenav/sidenav/sidenav.component.scss
@@ -34,7 +34,7 @@
     background-color: #ccc;
 }
 
-.vertical-menu a.active, .active-link {
+.vertical-menu a.active, .active-link, a.active-link-1 {
     padding-top: 11px !important;
     border-top: 4px solid #ffdd00;
     background-color: #dfdfdf;
